+++ tests/fixtures/simple/b/c.js
@@ -1,7 +1,6 @@
 module.exports = {
   command: function (args, flags, context) {
-    context.t.ok(flags.hi)
-    context.t.ok(args[0] = 'huh')
+    context.t.ok(flags.hi === 'huh')
   },
   options: []
 }