+++ tests/index.js
@@ -3,24 +3,6 @@ var test = require('tape')
 
 var directoryCommand = require('../index')
 
-test('no command found, default used', function (t) {
-  var directory = path.join(__dirname, 'fixtures', 'simple')
-
-  var defaultCommand = {
-    command: function (args, flags) {
-      t.ok(flags.hi)
-      t.end()
-    },
-    options: []
-  }
-
-  var options = {
-    defaultCommand
-  }
-
-  directoryCommand(directory, ['d', '--hi'], options)
-})
-
 test('simple usage', function (t) {
   var directory = path.join(__dirname, 'fixtures', 'simple')
 
@@ -33,7 +15,11 @@ test('simple usage', function (t) {
   directoryCommand(directory, ['a', '--hi', 'huh'], options, (a, b, c) => {
     directoryCommand(directory, ['b', '--hi', 'huh'], options, (a, b, c) => {
       directoryCommand(directory, ['b c', '--hi', 'huh'], options, (a, b, c) => {
-        t.end()
+        directoryCommand(directory, ['', '--hi', 'huh'], options, (a, b, c) => {
+          directoryCommand(directory, ['d', '--hi', 'huh'], options, (a, b, c) => {
+            t.end()
+          })
+        })
       })
     })
   })