+++ CHANGELOG.md
@@ -6,9 +6,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## Unreleased
 
-* ???
+- ... _nothing yet_
 
-## v1.0.0
+## [v2.0.0] - 2018-07-22
+
+### Changed
+
+- Default command is now specified by creating an index.js file in the commands directory. See the [basic usage example](examples/basic-usage).
+
+## [v1.0.1] - 2018-07-01
+
+### Fixed
+- fixed argument order
+
+## [v1.0.0] - 2017-05-20
 
 ### Added
 
@@ -35,4 +46,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   directoryCommand(directory, process.argv.slice(2), options)
   ```
 
+[v2.0.0]: https://github.com/sethvincent/directory-command/tree/v1.0.1...v2.0.0
+[v1.0.1]: https://github.com/sethvincent/directory-command/tree/v1.0.0...v1.0.1
 [v1.0.0]: https://github.com/sethvincent/directory-command/tree/v1.0.0